+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## (Unreleased)
 
-- Add `rd_auto` attribute to `iosxe_vrf` resource and data sources
+- Add `ip_local_proxy_arp` attribute to `iosxe_interface_vlan` resource and data sources
+- Add `iosxe_crypto` resource and data source
+- Add `rd_auto` attribute to `iosxe_vrf` resource and data source
 
 ## 0.11.0